sort of but not quite

+++ kernel/feather_m0_rfm69_arduino/src/icedq/icedq.h
@@ -20,8 +20,8 @@ typedef struct{
 typedef struct{
 	volatile uint8_t* queue;
 	volatile uint32_t capacity;
-	volatile int32_t head;  //NOTE1: These 2 are how processes synchronize, so make them visible to
-	volatile int32_t tail;  //       each other with volatile i.e. tell the compiler to not optimize them
+	volatile int8_t head;  //NOTE1: These 2 are how processes synchronize, so make them visible to
+	volatile int8_t tail;  //       each other with volatile i.e. tell the compiler to not optimize them
                           //NOTE2: These 2 NEED TO BE SIGNED for quere tail and head arithmetic
                           //       to work.
 }tIcedQQueue;
